PR #531 review feedback: tests for isTranslation, by-value SVGDocument, shared MockRendererInterface
Three review items from #531 landed together since they touch overlapping
files:

- **`isTranslation()` + `translation()` tests** (review comment on
  `donner/base/Transform.h`). Added `Transform.IsTranslation` and
  `Transform.Translation` to `Transform_tests.cc`: identity-case,
  pure-translation, non-translation (scale/rotate/skew/compose),
  floating-point jitter tolerance, and the rotated-then-translated
  case where the accessor still returns the matrix's `(e, f)` but
  `isTranslation()` correctly returns false.
- **SVGDocument stored by value** (review comment on
  `AsyncRenderer.h:249`). `compositorDocument_` was a raw pointer;
  changed to `std::optional<svg::SVGDocument>`. `SVGDocument` is a
  thin value facade over `std::shared_ptr<Registry>`, so copying is
  a refcount bump. Identity comparison now uses `handle().get()` so
  two facades wrapping the same Registry compare equal — which is
  the right "same document" semantic. Doc comment on the field
  spells out why.
- **MockRendererInterface dedupe** (review comment on
  `CompositorController_tests.cc:21`). The 35-line mock was copy-
  pasted across `CompositorController_tests.cc`,
  `DragSession_tests.cc`, `CompositorPerf_tests.cc`,
  `DualPathVerifier_tests.cc`, and `RendererDriver_tests.cc`.
  Promoted it to `donner/svg/renderer/tests/MockRendererInterface.h`
  + `:mock_renderer_interface` testonly library; the five consumers
  now `using MockRendererInterface = tests::MockRendererInterface;`
  to keep call-sites unchanged. One header, five deletions.

All five affected test suites pass. CMake-validate passes.
