+++ donner/svg/compositor/BUILD.bazel
@@ -67,6 +67,7 @@ donner_cc_test(
     deps = [
         ":compositor",
         "//donner/svg/renderer:renderer_interface",
+        "//donner/svg/renderer/tests:mock_renderer_interface",
         "//donner/svg/renderer:renderer_utils",
         "//donner/svg/tests:parser_test_utils",
         "@com_google_gtest//:gtest_main",
@@ -81,6 +82,7 @@ donner_cc_test(
     deps = [
         ":compositor",
         "//donner/svg/renderer:renderer_interface",
+        "//donner/svg/renderer/tests:mock_renderer_interface",
         "//donner/svg/tests:parser_test_utils",
         "@com_google_gtest//:gtest_main",
     ],
@@ -180,6 +182,7 @@ donner_cc_test(
         ":compositor",
         ":dual_path_verifier",
         "//donner/svg/renderer:renderer_interface",
+        "//donner/svg/renderer/tests:mock_renderer_interface",
         "//donner/svg/tests:parser_test_utils",
         "@com_google_gtest//:gtest_main",
     ],
@@ -195,6 +198,7 @@ donner_cc_test(
     deps = [
         ":compositor",
         "//donner/svg/renderer:renderer_interface",
+        "//donner/svg/renderer/tests:mock_renderer_interface",
         "//donner/svg/renderer:renderer_utils",
         "//donner/svg/tests:parser_test_utils",
         "@com_google_gtest//:gtest_main",
